Fix suggested changes.
