+++ game/common/string_helpers.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <string>
+
+// Helper functions for string-related things
+
+namespace sbat {
+  bool EndsWith(const std::string checked, const std::string suffix);
+}  // namespace sbat
