+++ game/common/win_helpers.cpp
@@ -131,13 +131,4 @@ wstring GetDocumentsPath() {
   return documents_path;
 }
 
-bool EndsWith(const string checked, const string suffix) {
-  if (suffix.length() > checked.length()) {
-    return false;
-  }
-
-  int index = checked.rfind(suffix);
-  return index != string::npos && (index + suffix.length() == checked.length());
-}
-
 }  // namespace sbat
