+++ game/common/win_helpers.h
@@ -92,6 +92,5 @@ private:
 };
 
 std::wstring GetDocumentsPath();
-bool EndsWith(const std::string checked, const std::string suffix);
 
 }  // namespace sbat
