+++ game/shieldbattery/shieldbattery.cpp
@@ -14,6 +14,7 @@
 
 #include "shieldbattery/settings.h"
 #include "common/func_hook.h"
+#include "common/string_helpers.h"
 #include "common/types.h"
 #include "common/win_helpers.h"
 #include "logger/logger.h"
@@ -140,7 +141,6 @@ NODE_EXTERN void QueueWorkForUiThread(void* arg,
   uv_mutex_unlock(&work_queue_mutex);
 }
 
-wstring userDataPath;
 void StartNode(void* arg) {
   HMODULE module_handle;
   wchar_t path[MAX_PATH];
@@ -161,7 +161,7 @@ void StartNode(void* arg) {
   assert(numProcessArgs >= 3);
   wchar_t* gameId = processArgs[0];
   wchar_t* port = processArgs[1];
-  userDataPath = processArgs[2];
+  wchar_t* userDataPath = processArgs[2];
 
   vector<wchar_t*> argv;
   argv.push_back(path);
@@ -172,7 +172,7 @@ void StartNode(void* arg) {
   argv.push_back(L"shieldbattery");
   argv.push_back(gameId);
   argv.push_back(port);
-  argv.push_back(&userDataPath[0]);
+  argv.push_back(userDataPath);
 
   // Convert argv to to UTF8
   vector<char*> utf8_argv;
@@ -388,15 +388,10 @@ DWORD __stdcall GetModuleFileNameAHook(
 
 HMODULE __stdcall LoadLibraryAHook(_In_ LPCTSTR filename) {
   if (EndsWith(filename, "local.dll")) {
-    wchar_t starcraft_dir[MAX_PATH];
-    DWORD count = GetCurrentDirectoryW(sizeof(starcraft_dir), starcraft_dir);
-    if (!count) {
-      ExitProcess(GetLastError());
-    }
-    wstring local_dll_path = wstring(starcraft_dir) + L"\\local.dll";
-    if (!PathFileExistsW(local_dll_path.c_str())) {
-      local_dll_path = userDataPath + L"\\downgrade\\local.dll";
-    }
+    wchar_t starcraft_path[MAX_PATH];
+    GetModuleFileNameW(NULL, starcraft_path, sizeof(starcraft_path));
+    wstring local_dll_path = wstring(starcraft_path);
+    local_dll_path.replace(local_dll_path.end() - strlen("starcraft.exe"), local_dll_path.end(), L"local.dll");
     return LoadLibraryW(local_dll_path.c_str());
   } else {
     return LoadLibraryA(filename);
